requirements.txt
langchain
python-dotenv
google-cloud-storage
google-cloud-bigquery
google-cloud-texttospeech
google-cloud-translate
google-cloud-vision
Flask
Pillow
beautifulsoup4
google-cloud-datastore
google-cloud-aiplatform